refactor: move key hints to GUIManager
